Fixes #2402: Restructure folders and move components into shared
