+++ src/__tests__/components/About/Overview.js
@@ -1,0 +1,9 @@
+import React from 'react';
+import Overview from '../../../components/About/Overview';
+import { shallow } from 'enzyme';
+
+describe('<Overview />', () => {
+  it('renders Overview without crashing', () => {
+    shallow(<Overview location={{ pathname: '/overview' }} />);
+  });
+});
