+++ src/__tests__/components/About/Support.js
@@ -1,16 +1,16 @@
 import React from 'react';
-import Logout from '../../../components/Auth/Logout';
+import Support from '../../../components/About/Support';
 import { shallow } from 'enzyme';
-import configureMockStore from 'redux-mock-store';
 import { Provider } from 'react-redux';
+import configureMockStore from 'redux-mock-store';
 const mockStore = configureMockStore();
 const store = mockStore({});
 
-describe('<Logout />', () => {
-  it('render Logout without crashing', () => {
+describe('<Support />', () => {
+  it('renders Support without crashing', () => {
     shallow(
       <Provider store={store}>
-        <Logout />
+        <Support location={{ pathname: '/support' }} />
       </Provider>,
     );
   });
