+++ src/__tests__/components/About/Team/SocialLinkButtons.js
@@ -1,0 +1,9 @@
+import React from 'react';
+import SocialLinkButtons from '../../../../components/About/Team/SocialLinkButtons';
+import { shallow } from 'enzyme';
+
+describe('<SocialLinkButtons />', () => {
+  it('render SocialLinkButtons without crashing', () => {
+    shallow(<SocialLinkButtons />);
+  });
+});
