+++ src/__tests__/components/NavigationBar/TopMenu.js
@@ -1,16 +1,16 @@
 import React from 'react';
-import Logout from '../../../components/Auth/Logout';
+import TopMenu from '../../../components/NavigationBar/TopMenu';
 import { shallow } from 'enzyme';
-import configureMockStore from 'redux-mock-store';
 import { Provider } from 'react-redux';
+import configureMockStore from 'redux-mock-store';
 const mockStore = configureMockStore();
 const store = mockStore({});
 
-describe('<Logout />', () => {
-  it('render Logout without crashing', () => {
+describe('<TopMenu />', () => {
+  it('renders TopMenu without crashing', () => {
     shallow(
       <Provider store={store}>
-        <Logout />
+        <TopMenu />
       </Provider>,
     );
   });
