--- src/__tests__/components/Overview/Overview.js
@@ -1,9 +1,0 @@
-import React from 'react';
-import Overview from '../../../components/Overview/Overview.react';
-import { shallow } from 'enzyme';
-
-describe('<Overview />', () => {
-  it('renders Overview without crashing', () => {
-    shallow(<Overview location={{ pathname: '/overview' }} />);
-  });
-});
