+++ src/__tests__/components/ProtectedRoute/index.js
@@ -1,16 +1,16 @@
 import React from 'react';
-import Logout from '../../../components/Auth/Logout';
+import ProtectedRoute from '../../../components/shared/ProtectedRoute';
 import { shallow } from 'enzyme';
-import configureMockStore from 'redux-mock-store';
 import { Provider } from 'react-redux';
+import configureMockStore from 'redux-mock-store';
 const mockStore = configureMockStore();
 const store = mockStore({});
 
-describe('<Logout />', () => {
-  it('render Logout without crashing', () => {
+describe('<ProtectedRoute />', () => {
+  it('render ProtectedRoute without crashing', () => {
     shallow(
       <Provider store={store}>
-        <Logout />
+        <ProtectedRoute />
       </Provider>,
     );
   });
