--- src/__tests__/components/ProtectedRoutes/ProtectedRoutes.js
@@ -1,17 +1,0 @@
-import React from 'react';
-import ProtectedRoute from '../../../components/ProtectedRoute';
-import { shallow } from 'enzyme';
-import { Provider } from 'react-redux';
-import configureMockStore from 'redux-mock-store';
-const mockStore = configureMockStore();
-const store = mockStore({});
-
-describe('<ProtectedRoute />', () => {
-  it('render ProtectedRoute without crashing', () => {
-    shallow(
-      <Provider store={store}>
-        <ProtectedRoute />
-      </Provider>,
-    );
-  });
-});
