--- src/__tests__/components/StaticAppBar/LeftMenu.js
@@ -1,16 +1,0 @@
-import React from 'react';
-import LeftMenu from '../../../components/StaticAppBar/LeftMenu';
-import { shallow } from 'enzyme';
-import { Provider } from 'react-redux';
-import configureMockStore from 'redux-mock-store';
-const mockStore = configureMockStore();
-const store = mockStore({});
-describe('<LeftMenu />', () => {
-  it('renders LeftMenu without crashing', () => {
-    shallow(
-      <Provider store={store}>
-        <LeftMenu />
-      </Provider>,
-    );
-  });
-});
