--- src/__tests__/components/StaticAppBar/TopMenu.js
@@ -1,17 +1,0 @@
-import React from 'react';
-import TopMenu from '../../../components/StaticAppBar/TopMenu';
-import { shallow } from 'enzyme';
-import { Provider } from 'react-redux';
-import configureMockStore from 'redux-mock-store';
-const mockStore = configureMockStore();
-const store = mockStore({});
-
-describe('<TopMenu />', () => {
-  it('renders TopMenu without crashing', () => {
-    shallow(
-      <Provider store={store}>
-        <TopMenu />
-      </Provider>,
-    );
-  });
-});
