--- src/__tests__/components/Team/SocialLinkButtons.js
@@ -1,9 +1,0 @@
-import React from 'react';
-import SocialLinkButtons from '../../../components/Team/SocialLinkButtons';
-import { shallow } from 'enzyme';
-
-describe('<SocialLinkButtons />', () => {
-  it('render SocialLinkButtons without crashing', () => {
-    shallow(<SocialLinkButtons />);
-  });
-});
