+++ src/components/NavigationBar/LeftMenu.js
