+++ src/components/cms/AuthorSkills/AuthorSkills.js
@@ -14,7 +14,7 @@ import skillActions from '../../../redux/actions/skill';
 import uiActions from '../../../redux/actions/ui';
 import Img from 'react-image';
 import ISO6391 from 'iso-639-1';
-import CircleImage from '../../CircleImage/CircleImage';
+import CircleImage from '../../shared/CircleImage';
 import { urls } from '../../../utils';
 import githubLogo from '../../../images/github-logo.png';
 import CloseButton from '../../shared/CloseButton';
