--- src/components/cms/CircleImage/CircleImage.css
@@ -1,19 +1,0 @@
-.CircleImage {
-    display: inline-block;
-    color: white;
-}
-
-.CircleImage--inner {
-    box-sizing: border-box;
-    white-space: nowrap;
-    overflow: hidden;
-    text-overflow: ellipsis;
-}
-
-.CircleImage--light {
-    color: gray;
-}
-
-.CircleImage--light .CircleImage--inner {
-    border: 1px solid lightgray;
-}
